Refactoring to a base class
to cut down on code duplication
